[feat] 플레이어 스태미나 추가

+++ Assets/Scripts/Data/GameData.cs
@@ -90,7 +90,7 @@ public class Skill : GameDataBase
 {
     public string SkillName;
     public int ATK;
-    public int Stamina;
+    public float Stamina;
     public int Level;
     public string Type;
     public string Anim;
